+++ src/main/java/racingcar/Main.java
@@ -1,6 +1,9 @@
 package racingcar;
 
 import java.util.Scanner;
+import racingcar.controller.RacingCarController;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class Main {
