+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -1,4 +1,9 @@
-package racingcar;
+package racingcar.controller;
+
+import racingcar.domain.RacingCars;
+import racingcar.domain.RacingGame;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class RacingCarController {
