+++ src/test/java/racingcar/controller/RacingCarControllerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 public class RacingCarControllerTest {
