+++ src/test/java/racingcar/view/OutputViewTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
